+++ app/Livewire/CamposUbicacionPropiedad.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class CamposUbicacionPropiedad extends Component
+{
+    public function render()
+    {
+        return view('livewire.campos-ubicacion-propiedad');
+    }
+}
