+++ app/Livewire/InformacionPropiedad.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class InformacionPropiedad extends Component
+{
+    public function render()
+    {
+        return view('livewire.informacion-propiedad');
+    }
+}
