+++ app/Livewire/UbicacionMapa.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class UbicacionMapa extends Component
+{
+    public function render()
+    {
+        return view('livewire.ubicacion-mapa');
+    }
+}
