+++ package.json
@@ -10,6 +10,7 @@
         "axios": "^1.7.4",
         "concurrently": "^9.0.1",
         "laravel-vite-plugin": "^1.0",
+        "leaflet": "^1.9.4",
         "postcss": "^8.4.47",
         "tailwindcss": "^3.4.15",
         "vite": "^5.0"
