+++ resources/views/livewire/informacion-propiedad.blade.php
@@ -1,0 +1,3 @@
+<div>
+    
+</div>
