+++ resources/views/livewire/registro-propiedad.blade.php
@@ -6,114 +6,23 @@
           <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
         </div>
         <div class="modal-body">
-            <div class="bg-white h-screen w-screen sm:px-8 md:px-16 sm:py-8 flex flex-row">
-                <main class="container mx-auto max-w-screen-lg h-full flex-1">
-                  <!-- file upload modal -->
-                  <article aria-label="File Upload Modal" class="relative h-full flex flex-col bg-white shadow-xl rounded-md" ondrop="dropHandler(event);" ondragover="dragOverHandler(event);" ondragleave="dragLeaveHandler(event);" ondragenter="dragEnterHandler(event);">
-                    <!-- overlay -->
-                    <div id="overlay" class="w-full h-full absolute top-0 left-0 pointer-events-none z-50 flex flex-col items-center justify-center rounded-md">
-                      <i>
-                        <svg class="fill-current w-12 h-12 mb-3 text-blue-700" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                          <path d="M19.479 10.092c-.212-3.951-3.473-7.092-7.479-7.092-4.005 0-7.267 3.141-7.479 7.092-2.57.463-4.521 2.706-4.521 5.408 0 3.037 2.463 5.5 5.5 5.5h13c3.037 0 5.5-2.463 5.5-5.5 0-2.702-1.951-4.945-4.521-5.408zm-7.479-1.092l4 4h-3v4h-2v-4h-3l4-4z" />
-                        </svg>
-                      </i>
-                      <p class="text-lg text-blue-700">Arrasta o Selecciona Imagenes</p>
-                    </div>
-            
-                    <!-- scroll area -->
-                    <section class="h-full overflow-auto p-8 w-full h-full flex flex-col">
-                      <header class="border-dashed border-2 border-gray-400 py-12 flex flex-col justify-center items-center">
-                        <p class="mb-3 font-semibold text-gray-900 flex flex-wrap justify-center">
-                          <span>Arrasta o Selecciona Imagenes</span>
-                        </p>
-                        <input id="hidden-input" type="file" multiple class="hidden" />
-                        <button id="button" class="mt-2 rounded-sm px-3 py-1 bg-gray-200 hover:bg-gray-300 focus:shadow-outline focus:outline-none">
-                          Seleccionar Archivos
-                        </button>
-                      </header>
-            
-                      <h1 class="pt-8 pb-3 font-semibold sm:text-lg text-gray-900">
-                        Carga Imagen
-                      </h1>
-            
-                      <ul id="gallery" class="flex flex-1 flex-wrap -m-1">
-                        <li id="empty" class="h-full w-full text-center flex flex-col items-center justify-center items-center">
-                          <img class="mx-auto w-32" src="https://user-images.githubusercontent.com/507615/54591670-ac0a0180-4a65-11e9-846c-e55ffce0fe7b.png" alt="no data" />
-                          <span class="text-small text-gray-500">No files selected</span>
-                        </li>
-                      </ul>
-                    </section>
-            
-                  </article>
-                </main>
-                    
-                @livewire('tipo-campos-registro-propiedad')
-
-               
+          <form>
+            <div class="flex flex-row flex-wrap">
+              @livewire('subida-archivo', ['tipoSubidaArchivo' => 'imagenes', 'widthProperty' => 'w-50', 'heightProperty' => 'h-full', 'mensaje' => 'Arrasta o Selecciona Imagenes del nuevo proyecto', 'subtitulo' => 'Carga Imagen Proyecto'])
+              @livewire('tipo-campos-registro-propiedad')
+              <div class="w-100">
+                <h2 class="fw-bold ps-2 fs-4">Estructura y Videos del Proyecto</h2>
+                <div class="flex flex-row">
+                  @livewire('subida-archivo', ['tipoSubidaArchivo' => 'planos', 'widthProperty' => 'w-100', 'heightProperty' => 'h-3/4', 'mensaje' => 'Ingresa los planos del proyecto', 'subtitulo' => 'Carga Planos Proyecto'])
+                  @livewire('subida-archivo', ['tipoSubidaArchivo' => 'videos', 'widthProperty' => 'w-100', 'heightProperty' => 'h-3/4', 'mensaje' => 'Arrasta o Suelta los videos del proyecto', 'subtitulo' => 'Carga Videos Proyecto'])
+                </div>
+              </div>
+              <div class="w-100">
+                @livewire('campos-ubicacion-propiedad')
+              </div>
             </div>
-
-            <div class="recursos-adicionales bg-white h-screen w-screen sm:px-8 md:px-16 sm:py-8 flex flex-row">
-                <h2 class="fw-bold text-2xl">Recursos Adicionales del Proyecto</h2>
-
-                <div></div>
-            </div>
-            
-              <!-- using two similar templates for simplicity in js code -->
-            <template id="file-template">
-                <li class="block p-1 w-1/2 sm:w-1/3 md:w-1/4 lg:w-1/6 xl:w-1/8 h-24">
-                  <article tabindex="0" class="group w-full h-full rounded-md focus:outline-none focus:shadow-outline elative bg-gray-100 cursor-pointer relative shadow-sm">
-                    <img alt="upload preview" class="img-preview hidden w-full h-full sticky object-cover rounded-md bg-fixed" />
-            
-                    <section class="flex flex-col rounded-md text-xs break-words w-full h-full z-20 absolute top-0 py-2 px-3">
-                      <h1 class="flex-1 group-hover:text-blue-800"></h1>
-                      <div class="flex">
-                        <span class="p-1 text-blue-800">
-                          <i>
-                            <svg class="fill-current w-4 h-4 ml-auto pt-1" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                              <path d="M15 2v5h5v15h-16v-20h11zm1-2h-14v24h20v-18l-6-6z" />
-                            </svg>
-                          </i>
-                        </span>
-                        <p class="p-1 size text-xs text-gray-700"></p>
-                        <button class="delete ml-auto focus:outline-none hover:bg-gray-300 p-1 rounded-md text-gray-800">
-                          <svg class="pointer-events-none fill-current w-4 h-4 ml-auto" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                            <path class="pointer-events-none" d="M3 6l3 18h12l3-18h-18zm19-4v2h-20v-2h5.711c.9 0 1.631-1.099 1.631-2h5.316c0 .901.73 2 1.631 2h5.711z" />
-                          </svg>
-                        </button>
-                      </div>
-                    </section>
-                  </article>
-                </li>
-            </template>
-            
-            <template id="image-template">
-                <li class="block p-1 w-1/2 sm:w-1/3 md:w-1/4 lg:w-1/6 xl:w-1/8 h-24">
-                  <article tabindex="0" class="group hasImage w-full h-full rounded-md focus:outline-none focus:shadow-outline bg-gray-100 cursor-pointer relative text-transparent hover:text-white shadow-sm">
-                    <img alt="upload preview" class="img-preview w-full h-full sticky object-cover rounded-md bg-fixed" />
-            
-                    <section class="flex flex-col rounded-md text-xs break-words w-full h-full z-20 absolute top-0 py-2 px-3">
-                      <h1 class="flex-1"></h1>
-                      <div class="flex">
-                        <span class="p-1">
-                          <i>
-                            <svg class="fill-current w-4 h-4 ml-auto pt-" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                              <path d="M5 8.5c0-.828.672-1.5 1.5-1.5s1.5.672 1.5 1.5c0 .829-.672 1.5-1.5 1.5s-1.5-.671-1.5-1.5zm9 .5l-2.519 4-2.481-1.96-4 5.96h14l-5-8zm8-4v14h-20v-14h20zm2-2h-24v18h24v-18z" />
-                            </svg>
-                          </i>
-                        </span>
-            
-                        <p class="p-1 size text-xs"></p>
-                        <button class="delete ml-auto focus:outline-none hover:bg-gray-300 p-1 rounded-md">
-                          <svg class="pointer-events-none fill-current w-4 h-4 ml-auto" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                            <path class="pointer-events-none" d="M3 6l3 18h12l3-18h-18zm19-4v2h-20v-2h5.711c.9 0 1.631-1.099 1.631-2h5.316c0 .901.73 2 1.631 2h5.711z" />
-                          </svg>
-                        </button>
-                      </div>
-                    </section>
-                  </article>
-                </li>
-              </template>
-            
+            <input type="submit" class="btn btn-registrar text-white mt-4 fs-5 fw-300" value="👉 Registrar Propiedad">
+          </form>
         </div>
         {{-- <div class="modal-footer">
           <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
