+++ resources/views/livewire/tipo-campos-registro-propiedad.blade.php
@@ -1,9 +1,9 @@
-<div class="formularioRegistroPropiedad flex-2 w-50">
+<div class="formularioRegistroPropiedad w-50">
     <h2 class="text-xl fw-bold">Tipo Proyecto</h2>
     <div class="col">
         <select class="form-select" aria-label="Default select example" wire:model.live="typeProjects">
-            <option selected>Selecciona tipo de proyecto</option>
-            <option value="1">
+            <option value="0" disabled>Selecciona tipo de proyecto</option>
+            <option value="1" selected>
                 🏢 Departamento
             </option>
             <option value="2" id="registro-casa">
@@ -15,7 +15,7 @@
         </select>
     </div>
 
-    <div class="campos mt-5">
+    <div class="campos mt-2">
   
         @if (intval($typeProjects) === 1 || intval($typeProjects) === 2)
             <h3 class="fw-bold pt-2 pb-2">Datos Generales Proyecto</h3>
@@ -60,7 +60,7 @@
                 @else
                     <div class="form-floating mb-3  flex-1">
                         <input type="text" class="form-control" id="floatingInput" placeholder="Inserta la profundidad del terreno">
-                        <label for="floatingInput">>📐 Profundidad Terreno</label>
+                        <label for="floatingInput">📐 Profundidad Terreno</label>
                     </div>
                 @endif 
             </div>
@@ -85,6 +85,9 @@
                     <label for="floatingInput">💲💲 Ponle un precio al proyecto</label>
                 </div>
             </div>
+
+            <h3 class="fw-bold pt-0 pb-2">Descripcion</h3>
+            <textarea id="message" rows="4" class="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-white-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-black dark:focus:ring-blue-500 dark:focus:border-blue-500" placeholder="Your message..."></textarea>
         @endif
     </div>
 </div>
