+++ resources/views/livewire/ubicacion-mapa.blade.php
@@ -1,0 +1,8 @@
+<div id="map" style="height:25rem;">
+   
+</div>
+
+
+<script>
+    var map = L.map('map').setView([51.505, -0.09], 13);
+</script>
